pyproject.toml
[tool.poetry]
name = "hw-27"
version = "0.1.0"
description = ""
authors = ["Deadpo_Ol <https://github.com/devoind>"]
readme = "README.md"
packages = [{ include = "hw_27" }]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
psycopg2 = "^2.9.4"
Pillow = "^9.2.0"
djangorestframework = "^3.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
